Development

+++ pantry-service/src/main/java/com/fcastro/pantry/product/ProductService.java
@@ -1,7 +1,11 @@
 package com.fcastro.pantry.product;
 
+import com.fcastro.pantry.exception.DatabaseConstraintException;
+import com.fcastro.pantry.exception.RequestParamExpectedException;
 import com.fcastro.pantry.exception.ResourceNotFoundException;
+import com.fcastro.pantry.pantryItem.PantryItemRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,11 +16,13 @@ import java.util.stream.Collectors;
 public class ProductService {
 
     private final ProductRepository repository;
+    private final PantryItemRepository pantryItemRepository;
     private final ModelMapper modelMapper;
 
 
-    public ProductService(ProductRepository repository, ModelMapper modelMapper) {
+    public ProductService(ProductRepository repository, PantryItemRepository pantryItemRepository, ModelMapper modelMapper) {
         this.repository = repository;
+        this.pantryItemRepository = pantryItemRepository;
         this.modelMapper = modelMapper;
     }
 
@@ -26,8 +32,17 @@ public class ProductService {
     }
 
     //TODO: Pageable
+    public List<ProductDto> getAll(String searchParam) {
+        List<Product> listEntity;
+
+        if (searchParam == null) throw new RequestParamExpectedException("Expecting to receive SearchParam: code or description value");
+
+        listEntity = repository.findAllByCodeOrDescription(searchParam.toLowerCase());
+        return listEntity.stream().map(this::convertToDTO).collect(Collectors.toList());
+    }
+
     public List<ProductDto> getAll() {
-        var listEntity = repository.findAll();
+        List<Product> listEntity = repository.findAll(Sort.by("code"));
         return listEntity.stream().map(this::convertToDTO).collect(Collectors.toList());
     }
 
@@ -40,6 +55,9 @@ public class ProductService {
         repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Product not found"));
 
+        if (pantryItemRepository.countPantryItem(id) > 0)
+            throw new DatabaseConstraintException("Product can not be removed. It is referred in one or more pantry items.");
+
         repository.deleteById(id);
     }
 
@@ -50,6 +68,8 @@ public class ProductService {
 
     private Product convertToEntity(ProductDto dto) {
         if (dto == null) return null;
-        return modelMapper.map(dto, Product.class);
+        var entity = modelMapper.map(dto, Product.class);
+        entity.setCode(entity.getCode().toUpperCase());
+        return entity;
     }
 }
